Refactored recent sparse dot product code - Moved into DSCC package following the (undocumented) rule that mixed dense and sparse goes into sparse package BenchmarkMatrixVectorMult_DDRM - Made default configuration less CPU intensive so that runtime regression won't be slow.
