+++ main/ejml-dsparse/benchmarks/src/org/ejml/sparse/csc/mult/BenchmarkMatrixVectorMult_DDRM.java
@@ -21,19 +21,8 @@ import org.ejml.data.DMatrixRMaj;
 import org.ejml.data.DMatrixSparse;
 import org.ejml.data.DMatrixSparseCSC;
 import org.ejml.dense.row.RandomMatrices_DDRM;
-import org.ejml.dense.row.mult.MatrixVectorMult_DDRM;
 import org.ejml.sparse.csc.RandomMatrices_DSCC;
-import org.openjdk.jmh.annotations.Benchmark;
-import org.openjdk.jmh.annotations.BenchmarkMode;
-import org.openjdk.jmh.annotations.Fork;
-import org.openjdk.jmh.annotations.Measurement;
-import org.openjdk.jmh.annotations.Mode;
-import org.openjdk.jmh.annotations.OutputTimeUnit;
-import org.openjdk.jmh.annotations.Param;
-import org.openjdk.jmh.annotations.Scope;
-import org.openjdk.jmh.annotations.Setup;
-import org.openjdk.jmh.annotations.State;
-import org.openjdk.jmh.annotations.Warmup;
+import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
 import org.openjdk.jmh.runner.options.Options;
@@ -48,12 +37,14 @@ import java.util.concurrent.TimeUnit;
 @Measurement(iterations = 5)
 @State(Scope.Benchmark)
 @Fork(value = 1)
-public class BenchmarkMatrixVectorMult_DDRM_InnerProduct {
+public class BenchmarkMatrixVectorMult_DDRM {
 
-    @Param({"5", "1000", "3000"})
+    //    @Param({"5", "1000", "3000"})
+    @Param({"3000"})
     public int size;
 
-    @Param({"0.2", "0.4", "0.6", "0.8"})
+    //    @Param({"0.2", "0.4", "0.6", "0.8"})
+    @Param({"0.2"})
     public double density;
 
     private DMatrixSparse A;
@@ -64,19 +55,21 @@ public class BenchmarkMatrixVectorMult_DDRM_InnerProduct {
     public void setUp() {
         Random rand = new Random(234);
         B = RandomMatrices_DDRM.symmetricPosDef(size, rand);
-        A = RandomMatrices_DSCC.rectangle(size, 1, (int) (density * size), rand);
+        A = RandomMatrices_DSCC.rectangle(size, 1, (int)(density*size), rand);
         a = new DMatrixRMaj(A).data;
     }
 
-    @Benchmark public void innerProduct() { MatrixVectorMult_DDRM.innerProduct(a, 0, B, a, 0);}
-    @Benchmark public void innerProduct_sparse() { MatrixVectorMult_DDRM.innerProduct(A, B, A);}
-    @Benchmark public void innerProduct_sparse_csc() { MatrixVectorMult_DDRM.innerProduct((DMatrixSparseCSC) A, B, (DMatrixSparseCSC) A);}
-    @Benchmark public void innerProduct_symmetric_sparse() { MatrixVectorMult_DDRM.innerProductSelfSymmetrical(A, B);}
-    @Benchmark public void innerProduct_symmetric_sparse_csc() { MatrixVectorMult_DDRM.innerProductSelfSymmetrical((DMatrixSparseCSC) A, B);}
+    // @formatter:off
+//    @Benchmark public void innerProduct_dense() { MatrixVectorMult_DDRM.innerProduct(a, 0, B, a, 0);} // can be very slow
+    @Benchmark public void innerProduct_sparse() { MatrixVectorMult_DSCC.innerProduct(A, B, A);}
+    @Benchmark public void innerProduct_sparse_csc() { MatrixVectorMult_DSCC.innerProduct((DMatrixSparseCSC) A, B, (DMatrixSparseCSC) A);}
+    @Benchmark public void innerProduct_symmetric_sparse() { MatrixVectorMult_DSCC.innerProductSelfSymmetrical(A, B);}
+    @Benchmark public void innerProduct_symmetric_sparse_csc() { MatrixVectorMult_DSCC.innerProductSelfSymmetrical((DMatrixSparseCSC) A, B);}
+    // @formatter:on
 
     public static void main( String[] args ) throws RunnerException {
         Options opt = new OptionsBuilder()
-                .include(BenchmarkMatrixVectorMult_DDRM_InnerProduct.class.getSimpleName())
+                .include(BenchmarkMatrixVectorMult_DDRM.class.getSimpleName())
                 .build();
         new Runner(opt).run();
     }
